Updated Doxygen Comments

+++ Core/Inc/gpio.h
@@ -1,5 +1,11 @@
 #include "stm32f407xx.h"
 
-void static inline Toggle_Pin(GPIO_TypeDef* GPIO, int pin) {
+/**
+ * @brief Toggle a given GPIO pin on or off
+ * 
+ * @param GPIO [GPIO_TypeDef*] GPIO Port to use
+ * @param pin [uint8_t] Pin to toggle
+ */
+void static inline Toggle_Pin(GPIO_TypeDef* GPIO, uint8_t pin) {
   GPIO->ODR ^= (1 << pin);
 }
